Adjust inherited method parameter type

+++ src/main/java/io/smilo/api/peer/payloadhandler/sport/NetworkState.java
@@ -88,7 +88,7 @@ public class NetworkState implements INetworkState {
         return topBlock;
     }
 
-    public void setTopBlock(int topBlock) {
+    public void setTopBlock(long topBlock) {
         this.topBlock = topBlock;
         updateCatchupMode();
     }
